fix: Dash transferred out of the Dash Wallet after the transaction is confirmed
