+++ integrations/coinbase-integration/src/main/java/org/dash/wallet/integration/coinbase_integration/ui/CoinbaseConvertCryptoFragment.kt
@@ -36,7 +36,6 @@ import org.dash.wallet.common.ui.FancyAlertDialog
 import org.dash.wallet.common.ui.viewBinding
 import org.dash.wallet.common.util.GenericUtils
 import org.dash.wallet.common.util.safeNavigate
-import org.dash.wallet.integration.coinbase_integration.DASH_CURRENCY
 import org.dash.wallet.integration.coinbase_integration.R
 import org.dash.wallet.integration.coinbase_integration.databinding.FragmentCoinbaseConvertCryptoBinding
 import org.dash.wallet.integration.coinbase_integration.model.CoinBaseUserAccountDataUIModel
@@ -103,21 +102,21 @@ class CoinbaseConvertCryptoFragment : Fragment(R.layout.fragment_coinbase_conver
         convertViewModel.onContinueEvent.observe(viewLifecycleOwner) { pair ->
             val swapValueErrorType = convertViewModel.checkEnteredAmountValue()
             if (swapValueErrorType == SwapValueErrorType.NOError) {
-                if (!pair.first && selectedCoinBaseAccount?.coinBaseUserAccountData?.currency?.code != DASH_CURRENCY) {
-                    selectedCoinBaseAccount?.let {
-                        pair.second?.first?.let { fait ->
-                            viewModel.swapTrade(fait, it, pair.first)
-                        }
-                    }
-                } else {
-                    pair.second?.second?.let { coin ->
-                        selectedCoinBaseAccount?.let {
-                            pair.second?.first?.let { fait ->
-                                viewModel.sellDashToCoinBase(coin, fait, it)
-                            }
-                        }
+                //  if (!pair.first && selectedCoinBaseAccount?.coinBaseUserAccountData?.currency?.code != DASH_CURRENCY) {
+                selectedCoinBaseAccount?.let {
+                    pair.second?.first?.let { fait ->
+                        viewModel.swapTrade(fait, it, pair.first)
                     }
                 }
+//                } else {
+//                    pair.second?.second?.let { coin ->
+//                        selectedCoinBaseAccount?.let {
+//                            pair.second?.first?.let { fait ->
+//                                viewModel.sellDashToCoinBase(coin, fait, it)
+//                            }
+//                        }
+//                    }
+//                }
             } else {
                 showSwapValueErrorView(swapValueErrorType)
             }
@@ -133,49 +132,6 @@ class CoinbaseConvertCryptoFragment : Fragment(R.layout.fragment_coinbase_conver
             }
         )
 
-
-        viewModel.getUserAccountAddressFailedCallback.observe(viewLifecycleOwner) {
-            val placeBuyOrderError = CoinbaseGenericErrorUIModel(
-                R.string.error,
-                getString(R.string.error),
-                R.drawable.ic_info_red,
-                negativeButtonText = R.string.close
-            )
-            safeNavigate(
-                CoinbaseServicesFragmentDirections.coinbaseServicesToError(
-                    placeBuyOrderError
-                )
-            )
-        }
-
-
-        viewModel.onInsufficientMoneyCallback.observe(viewLifecycleOwner) {
-            val placeBuyOrderError = CoinbaseGenericErrorUIModel(
-                R.string.insufficient_money_title,
-                getString(R.string.insufficient_money_msg),
-                R.drawable.ic_info_red,
-                negativeButtonText = R.string.close
-            )
-            safeNavigate(
-                CoinbaseServicesFragmentDirections.coinbaseServicesToError(
-                    placeBuyOrderError
-                )
-            )
-        }
-
-        viewModel.onFailure.observe(viewLifecycleOwner) {
-            val placeBuyOrderError = CoinbaseGenericErrorUIModel(
-                R.string.send_coins_error_msg,
-                getString(R.string.insufficient_money_msg),
-                R.drawable.ic_info_red,
-                negativeButtonText = R.string.close
-            )
-            safeNavigate(
-                CoinbaseServicesFragmentDirections.coinbaseServicesToError(
-                    placeBuyOrderError
-                )
-            )
-        }
 
         viewModel.swapTradeFailedCallback.observe(viewLifecycleOwner) {
             val placeBuyOrderError = CoinbaseGenericErrorUIModel(
