+++ integrations/coinbase-integration/src/main/java/org/dash/wallet/integration/coinbase_integration/viewmodels/CoinbaseConversionPreviewViewModel.kt
@@ -23,8 +23,11 @@ import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
+import org.bitcoinj.core.Coin
+import org.bitcoinj.core.InsufficientMoneyException
 import org.dash.wallet.common.WalletDataProvider
 import org.dash.wallet.common.data.SingleLiveEvent
+import org.dash.wallet.common.services.SendPaymentService
 import org.dash.wallet.integration.coinbase_integration.model.*
 import org.dash.wallet.integration.coinbase_integration.network.ResponseResource
 import org.dash.wallet.integration.coinbase_integration.repository.CoinBaseRepositoryInt
@@ -34,7 +37,8 @@ import javax.inject.Inject
 @HiltViewModel
 class CoinbaseConversionPreviewViewModel @Inject constructor(
     private val coinBaseRepository: CoinBaseRepositoryInt,
-    private val walletDataProvider: WalletDataProvider
+    private val walletDataProvider: WalletDataProvider,
+    private val sendPaymentService: SendPaymentService
 ) : ViewModel() {
     private val _showLoading: MutableLiveData<Boolean> = MutableLiveData()
     val showLoading: LiveData<Boolean>
@@ -58,6 +62,11 @@ class CoinbaseConversionPreviewViewModel @Inject constructor(
 
     val swapTradeFailedCallback = SingleLiveEvent<String>()
 
+    val getUserAccountAddressFailedCallback = SingleLiveEvent<Unit>()
+    val sendDashToCoinBaseFailed = SingleLiveEvent<Unit>()
+    val onFailure = SingleLiveEvent<String>()
+    val onInsufficientMoneyCallback = SingleLiveEvent<Unit>()
+
     fun commitSwapTrade(params: SwapTradeUIModel) = viewModelScope.launch(Dispatchers.Main) {
         _showLoading.value = true
         when (val result = coinBaseRepository.commitSwapTrade(params.swapTradeId)) {
@@ -67,7 +76,12 @@ class CoinbaseConversionPreviewViewModel @Inject constructor(
                     commitBuyOrderFailedCallback.call()
                 } else {
                     if (params.inputCurrencyName.lowercase() == "dash") {
-                        _transactionCompleted.value = TransactionState(true, null)
+                        try {
+                            val coin = Coin.parseCoin(params.inputAmount.toString())
+                            sellDashToCoinBase(coin)
+                        } catch (x: Exception) {
+                            Coin.ZERO
+                        }
                     } else {
                         sendFundToWalletParams = SendTransactionToWalletParams(
                             amount = result.value.displayInputAmount,
@@ -170,5 +184,46 @@ class CoinbaseConversionPreviewViewModel @Inject constructor(
 
     fun onRefreshOrderClicked(swapTradeUIModel: SwapTradeUIModel) {
         swapTrade(swapTradeUIModel)
+    }
+
+    fun sellDashToCoinBase(coin: Coin) = viewModelScope.launch(Dispatchers.Main) {
+        _showLoading.value = true
+
+
+        when (val result = coinBaseRepository.createAddress()) {
+            is ResponseResource.Success -> {
+                if (result.value?.isEmpty() == true) {
+                    _showLoading.value = false
+                    getUserAccountAddressFailedCallback.call()
+                } else {
+                    result.value?.let {
+                        sendDashToCoinbase(coin, result.value)
+                        _transactionCompleted.value = TransactionState(true, null)
+                        _showLoading.value = false
+                    }
+                    _showLoading.value = false
+                }
+            }
+            is ResponseResource.Failure -> {
+                _showLoading.value = false
+                getUserAccountAddressFailedCallback.call()
+            }
+        }
+    }
+
+    private suspend fun sendDashToCoinbase(coin: Coin, addressInfo: String): Boolean {
+        val address = walletDataProvider.createSentDashAddress(addressInfo)
+        try {
+            val transaction = sendPaymentService.sendCoins(address, coin)
+            return transaction.isPending
+        } catch (x: InsufficientMoneyException) {
+            onInsufficientMoneyCallback.call()
+            x.printStackTrace()
+            return false
+        } catch (ex: Exception) {
+            onFailure.value = ex.message
+            ex.printStackTrace()
+            return false
+        }
     }
 }
