+++ integrations/coinbase-integration/src/main/java/org/dash/wallet/integration/coinbase_integration/viewmodels/CoinbaseConvertCryptoViewModel.kt
@@ -21,27 +21,23 @@ import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import org.bitcoinj.core.Coin
-import org.bitcoinj.core.InsufficientMoneyException
 import org.bitcoinj.utils.Fiat
 import org.dash.wallet.common.Configuration
 import org.dash.wallet.common.WalletDataProvider
 import org.dash.wallet.common.data.SingleLiveEvent
 import org.dash.wallet.common.livedata.Event
-import org.dash.wallet.common.services.SendPaymentService
 import org.dash.wallet.common.util.GenericUtils
 import org.dash.wallet.integration.coinbase_integration.DASH_CURRENCY
 import org.dash.wallet.integration.coinbase_integration.model.*
 import org.dash.wallet.integration.coinbase_integration.network.ResponseResource
 import org.dash.wallet.integration.coinbase_integration.repository.CoinBaseRepositoryInt
-import java.util.*
 import javax.inject.Inject
 
 @HiltViewModel
 class CoinbaseConvertCryptoViewModel @Inject constructor(
     private val coinBaseRepository: CoinBaseRepositoryInt,
     val config: Configuration,
-    private val walletDataProvider: WalletDataProvider,
-    private val sendPaymentService: SendPaymentService
+    private val walletDataProvider: WalletDataProvider
 ) : ViewModel() {
     private val _userAccountsInfo: MutableLiveData<List<CoinBaseUserAccountDataUIModel>> = MutableLiveData()
     val userAccountsInfo: LiveData<List<CoinBaseUserAccountDataUIModel>>
@@ -73,10 +69,6 @@ class CoinbaseConvertCryptoViewModel @Inject constructor(
     val dashWalletBalance: LiveData<Coin>
         get() = this._dashWalletBalance
 
-    val getUserAccountAddressFailedCallback = SingleLiveEvent<Unit>()
-    val onFailure = SingleLiveEvent<String>()
-    val onInsufficientMoneyCallback = SingleLiveEvent<Unit>()
-    val sendDashToCoinBaseFailed = SingleLiveEvent<Unit>()
     init {
         setDashWalletBalance()
         getUserAccountInfo()
@@ -110,49 +102,7 @@ class CoinbaseConvertCryptoViewModel @Inject constructor(
     }
 
 
-    fun sellDashToCoinBase(coin: Coin, valueToConvert: Fiat, selectedCoinBaseAccount: CoinBaseUserAccountDataUIModel) = viewModelScope.launch(Dispatchers.Main) {
-        _showLoading.value = true
 
-
-        when (val result = coinBaseRepository.createAddress()) {
-            is ResponseResource.Success -> {
-                if (result.value?.isEmpty() == true) {
-                    _showLoading.value = false
-                    getUserAccountAddressFailedCallback.call()
-                } else {
-                    result.value?.let {
-                        if (sendDashToCoinbase(coin, result.value)) {
-                            swapTrade(valueToConvert, selectedCoinBaseAccount, true)
-                        } else {
-                            _showLoading.value = false
-                            sendDashToCoinBaseFailed.call()
-                        }
-                    }
-                    _showLoading.value = false
-                }
-            }
-            is ResponseResource.Failure -> {
-                _showLoading.value = false
-                getUserAccountAddressFailedCallback.call()
-            }
-        }
-    }
-
-    private suspend fun sendDashToCoinbase(coin: Coin, addressInfo: String): Boolean {
-        val address = walletDataProvider.createSentDashAddress(addressInfo)
-        try {
-            val transaction = sendPaymentService.sendCoins(address, coin)
-            return transaction.isPending
-        } catch (x: InsufficientMoneyException) {
-            onInsufficientMoneyCallback.call()
-            x.printStackTrace()
-            return false
-        } catch (ex: Exception) {
-            onFailure.value = ex.message
-            ex.printStackTrace()
-            return false
-        }
-    }
 
     fun swapTrade(valueToConvert: Fiat, selectedCoinBaseAccount: CoinBaseUserAccountDataUIModel, dashToCrypt: Boolean) = viewModelScope.launch(Dispatchers.Main) {
         _showLoading.value = true
